Update User model and enhance authorization setup
Enhanced the User model by adding additional fields and changing the type of 'id' field from Integer to Long. Also, introduced RoleName and UserStatus enums for maintaining specific user states and role types. Set up initial roles and a default admin user during application startup using SetupDataLoader. Enabled JPA Auditing to keep track of entity changes. Updated relevant classes to handle changes in the User model and authorization setup.

+++ src/main/java/uz/etc/etcfitness/config/ApplicationAuditAware.java
@@ -8,9 +8,9 @@ import uz.etc.etcfitness.user.User;
 
 import java.util.Optional;
 
-public class ApplicationAuditAware implements AuditorAware<Integer> {
+public class ApplicationAuditAware implements AuditorAware<Long> {
     @Override
-    public Optional<Integer> getCurrentAuditor() {
+    public Optional<Long> getCurrentAuditor() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if (authentication == null ||
